feat(python): chase buffers following original buffers
Chase buffers will turn hidden if the original buffer is turns
hidden, and will be displayed back when original buffer is
displayed again.

Created initial structure to track chase buffers. Now the original
buffer number will be set as a variable to the chase buffer so we
can unset the reference while closing the chase buffer while
focused.

Refs: #6 #7