feat(python): Execute python tests based on the cursor location
Now the run_file will execute tests based on the cursor location
on the bufffer.

If the cursor is outiside any test suite, will run all tests from
the module related to the file.

If the cursor is inside a test suite, will run all tests from the
suite.

If the cursor is inside a test method, will run only this test
method.

Refs: #12