fix: handle pip show with on_exit as uv was adding to stderr
When uv is installed, pip show outputs to stderr all the time.

This causes issues when using on_stderr as the we think the data
returned is an error when it's not.

As we don't really care about stderr for pip show, we can just
check the exit code in on_exit to determine if the command
is successful or not.

Refs: #23